made extensive changes to finish functionality. hardcoded the scenario 4 and scenario 5 tests for simplicity. the simulation can now be run at a minimum viable product level.

+++ WildfireSimulation/Models/RainEvent.cs
@@ -2,6 +2,6 @@
 {
     public class RainEvent : Event
     {
-        public int? Rainfall { get; set; }
+        public double? Rainfall { get; set; }
     }
 }
